fix unit of refractory period during evaluation
after restructuring (cf. task #1580), the requested unit during
evaluation is the unit of the feature, i.e. for tau_refrac it is
seconds and not anymore the unit of the technical parameter

note that for some parameters, e.g. most potentials, the unit of feature
and the technical parameter are identical and therefor no change is
needed

Change-Id: I20fa354f2709925a31ffd9cbc9bc8d717233449e