Re-sampling from variable to fixed time step in SimStHALContainer
The circuit level simulator which is ultimately utilized by
SimStHALContainer may employ a numeric method with an adaptive
stepsize [1] for transient simulations. While this is beneficial for
simulation time and allows for a more efficient representation of the
ODE solution, it is not compatible with some of the analysis method
employed in the calibration framework, `cake` [2].
In particular, the TraceAverager deals with equidistantly sampled data.
Further, numerical methods such as fitting routines, may behave
differently for equally and non-equally sampled data, thus devaluing
sim_denmem as a calibration testing tool.

Hence, a conversion of circuit level simulation data into a format
which is similar to data obtained from the hardware system is highly
beneficial.

This conversion is enabled by default and can be disabled using the
`resample_output` argument in SimStHALContainer.__init__

[1] http://en.wikipedia.org/wiki/Adaptive_stepsize
[2] https://brainscales-r.kip.uni-heidelberg.de/projects/cake


Change-Id: Ia62ea677457022318e43a1d97a3b71520aa6f3af