add tool to collect links to calibration files
Many tools, e.g. marocco, expect calibration files to be present
in a certain format, "w{waferId}-h{hicannId}.xml", and take the
directory for the XML files as an argument
(e.g. PyMarocco().calib_path). At the moment, cake can create multiple
variants of the calibration files, differing in, e.g., whether the
small or the large capacitance setting is used for the neurons.
Automated calibration often creates separate (sub-)directories for
the output for different HICANNs and wafers, which in turn makes
automated use of these files inconvenient.

The script in this change simplifies the collection of calibration
files in one directory by using symbolic links to the original files.
The naming scheme for the links is the format that is expected by
marocco.

Change-Id: Ib7f92401496c65016d163a73c63e3b0507e7d8f7