Add saving of parameter range to HFDStore
In the evaluation, parameter.name + '_range' (from v4_eval.py) is not
in the measurement config, thus it has to be taken from the config of
the CalibrationUnit and added to the DataFrame that is saved. This
allows plotting/data evaluation w/o having to load the config from an
extra file.

Change-Id: I9b098e9d3d5f3e5d85bc05e63086928f841f26b8